+++ NEWS.md
@@ -1,5 +1,9 @@
 # hpa.spatial.data (development version)
 
+* update LHN shape file to include all of Tasmania as a single LHN called 
+  "Tasmanian Health Service" as per the latest release of LHN identifiers from
+  the AIHW (<https://meteor.aihw.gov.au/content/786748>).
+
 * update LHN shape file to use LHNs for Western Australia from
   <https://catalogue.data.wa.gov.au/dataset/health-services-health-006>.